all stats now 32bits

+++ inc/bplib.h
@@ -134,16 +134,16 @@ typedef struct {
 
 /* Bundle Channel Statistics */
 typedef struct {
-    uint16_t    lost;           // storage or copy failure, unable to retrieve
-    uint16_t    expired;        // lifetime expired, deliberately removed
-    uint16_t    acknowledged;   // freed by custody signal
-    uint16_t    transmitted;    // sent (includes resends)
-    uint16_t    retransmitted;  // timedout and resent
-    uint16_t    received;       // bundles processed
-    uint16_t    delivered;      // payloads accepted
-    uint16_t    bundles;        // number of data bundles currently in storage
-    uint16_t    payloads;       // number of payloads currently in storage
-    uint16_t    records;        // number of dacs bundles currently in storage
+    uint32_t    lost;           // storage or copy failure, unable to retrieve
+    uint32_t    expired;        // lifetime expired, deliberately removed
+    uint32_t    acknowledged;   // freed by custody signal
+    uint32_t    transmitted;    // sent (includes resends)
+    uint32_t    retransmitted;  // timedout and resent
+    uint32_t    received;       // bundles processed
+    uint32_t    delivered;      // payloads accepted
+    uint32_t    bundles;        // number of data bundles currently in storage
+    uint32_t    payloads;       // number of payloads currently in storage
+    uint32_t    records;        // number of dacs bundles currently in storage
     uint32_t    active;         // number of slots in active table in use
 } bp_stats_t;
